move ParserException and error code map into errors.py
