+++ terminalone/errors.py
@@ -88,3 +88,21 @@ class LoginError(T1Error):
     def __str__(self):
         return repr('{}: {} -- {}'.format(self.code, self.message,
                                           self.credentials))
+
+
+class ParserException(Exception):
+    def __init__(self, caught):
+        self.caught = caught
+
+# Map known status.code responses to Exceptions. 'ok' signifies no exception,
+# so that is None. 'invalid' can have many errors and needs
+# an additional level of parsing, while the others can be instantiated directly.
+STATUS_CODES = {
+    'ok': None,
+    'invalid': True,
+    'not_found': NotFoundError,
+    'auth_required': AuthRequiredError,
+    'auth_error': AuthRequiredError,
+    'error': APIError,
+    'bad_request': ClientError,
+}
