--- terminalone/exceptions.py
@@ -1,3 +1,0 @@
-class ParserException(Exception):
-    def __init__(self, caught):
-        self.caught = caught
